fix: use int64_t instead of long long

+++ data_structures/leetcode/maths/palindrome_number.cc
@@ -6,7 +6,7 @@ bool isPalindrome(int x) {
 	if (x <= 10) return false;
 
 	// use long long to avoid overflow error
-	long long rev = 0;
+	int64_t rev = 0;
 	while (x > 0) {
 		rev = rev * 10 + x % 10;
 		x = x/10;
